+++ src/redis-cli.c
@@ -8017,7 +8017,7 @@ void bytesToHuman(char *s, long long n) {
 
 static void statMode(void) {
     redisReply *reply;
-    long aux, requests = 0;
+    long long aux, requests = 0;
     int i = 0;
 
     while(1) {
@@ -8039,14 +8039,14 @@ static void statMode(void) {
         /* Keys */
         aux = 0;
         for (j = 0; j < 20; j++) {
-            long k;
+            long long k;
 
             sprintf(buf,"db%d:keys",j);
             k = getLongInfoField(reply->str,buf);
-            if (k == LONG_MIN) continue;
+            if (k == LLONG_MIN) continue;
             aux += k;
         }
-        sprintf(buf,"%ld",aux);
+        sprintf(buf,"%lld",aux);
         printf("%-11s",buf);
 
         /* Used memory */
@@ -8056,23 +8056,23 @@ static void statMode(void) {
 
         /* Clients */
         aux = getLongInfoField(reply->str,"connected_clients");
-        sprintf(buf,"%ld",aux);
+        sprintf(buf,"%lld",aux);
         printf(" %-8s",buf);
 
         /* Blocked (BLPOPPING) Clients */
         aux = getLongInfoField(reply->str,"blocked_clients");
-        sprintf(buf,"%ld",aux);
+        sprintf(buf,"%lld",aux);
         printf("%-8s",buf);
 
         /* Requests */
         aux = getLongInfoField(reply->str,"total_commands_processed");
-        sprintf(buf,"%ld (+%ld)",aux,requests == 0 ? 0 : aux-requests);
+        sprintf(buf,"%lld (+%lld)",aux,requests == 0 ? 0 : aux-requests);
         printf("%-19s",buf);
         requests = aux;
 
         /* Connections */
         aux = getLongInfoField(reply->str,"total_connections_received");
-        sprintf(buf,"%ld",aux);
+        sprintf(buf,"%lld",aux);
         printf(" %-12s",buf);
 
         /* Children */
